[CSL-1081] Split Pos.Block.Core into smaller modules
